requirements.txt
TensorFlow==2.8.0
hydra-core==1.3.2
scikit-learn==1.3.2
protobuf==3.20.1
rich
wandb
pandas